+++ R/detindex.R
@@ -1,12 +1,14 @@
 #' @export
 detindex <-
-function(storage, verbose=TRUE)
+function(storage, mini=c("B","D","M"), pmini=1, verbose=TRUE)
 {
      #                          detindex
      #
-     # VALUE    List containing index of detguides and related files in storage directory
+     # VALUE    List containing index of detguides and related files in storage directory or of one set of minidetguides for p=pmini
      #
      # INPUT    storage     Quoted name of directory for storage of detguides. 
+     #          mini        Index includes (B)oth detguides and minidetguides, (D)etguides only, (M)inidetguides only 
+     #          pmini       p for minidetguides and related files; ignored if mini="D"
      #
      #          verbose     Logical. T causes printing of program ID before and after running.
      #
@@ -28,17 +30,45 @@ function(storage, verbose=TRUE)
 
      out<- dir(path=storage)
      out <- out[out != "max.created.txt"]     
+     out <- sort(as.numeric(out))
      ndirs <- length(out)
-     detguide <- parseguide <- parseguidesym <- rep(0,ndirs)
-     subdir <- paste(storage,out,sep="/")
-     for(i in 1:ndirs){
-          filenames <- dir(path=subdir[i])
-          if(any(filenames=="detguide.txt"))     detguide[i] <- 1
 
-          if(any(filenames=="parseguide.txt"))   parseguide[i] <- 1
-          if(any(filenames=="parseguidesym.txt"))parseguidesym[i] <- 1
-     }
-     index=data.frame(p=out, detguide, parseguide, parseguidesym)
+
+     if(mini=="B" | mini=="D"){
+          detguide <- parseguide <- parseguidesym <- rep(0,ndirs)
+
+          subdir <- paste(storage,out,sep="/")
+          for(i in 1:ndirs){
+               filenames <- dir(path=subdir[i])
+               if(any(filenames=="detguide.txt"))     detguide[i] <- 1
+               if(any(filenames=="parseguide.txt"))   parseguide[i] <- 1
+               if(any(filenames=="parseguidesym.txt"))parseguidesym[i] <- 1
+          }
+          index=data.frame(p=out, detguide, parseguide, parseguidesym)
+     }    # mini==B or D
+     #
+     if(mini=="B" | mini=="M"){
+          storage <- paste(storage,pmini, sep="/")
+          if(!file.exists(storage)) stop("No storage for this value of pmini")
+          out<- dir(path=storage)
+          ndirs <- length(out)
+
+          minidetguide <- parseguide <- parseguidesym <- rep(0,ndirs)
+
+          subdir <- paste(storage,out,sep="/")
+          for(i in 1:ndirs){
+               filenames <- dir(path=subdir[i])
+               if(any(filenames=="minidetguide.txt")) minidetguide[i] <- 1
+               if(any(filenames=="parseguide.htm"))   parseguide[i] <- 1
+               if(any(filenames=="parseguidesym.htm"))parseguidesym[i] <- 1
+          }
+          index2=data.frame(R1Cs=out, minidetguide, parseguide, parseguidesym)
+     }    #   mini=="B" or M
+     #
+     if(mini=="D")outlast <- list(Detguides=index, Call=MC) 
+     if(mini=="M")outlast <- list(Minidetguides=index2, Call=MC) 
+     if(mini=="B")outlast <- list(Detguides=index, Minidetguides=index2, Call=MC) 
+     #
      if(verbose) {
           print("", quote = FALSE)
           print("Finished running detindex", quote = FALSE)
@@ -47,5 +77,5 @@ function(storage, verbose=TRUE)
           print("", quote = FALSE)
           print("1 indicates file is present, 0 not present", quote=FALSE)
      }
-     list(index, Call=MC)
+     outlast
 }